mass change of namespaces to have prefix net.ivoa.vodsl
This was to stop polluting the net.ivoa.vodml namespace which should
probably belong to the official VODML implementations.

It also makes the naming internally more consistent as the xtext
machinery likes to use single prefix for everything and there was a
subtle NPE that was difficult to track down because of the mixture.

hopefully this closes #8 as the ReadMe should be more explicit about how
to build the standalone parser and the build should work so closes #6
and I think also closes #7 because that pull request is not really
compatible with the changes here, although in the same spirit.